@@ -15,18 +15,18 @@ import android.os.Build.VERSION;
 import android.os.Build.VERSION_CODES;
 import android.os.Handler;
 import android.os.PowerManager;
-import android.support.annotation.LayoutRes;
-import android.support.annotation.NonNull;
-import android.support.design.widget.AppBarLayout;
-import android.support.design.widget.AppBarLayout.LayoutParams;
-import android.support.design.widget.AppBarLayout.LayoutParams.ScrollFlags;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v4.widget.SwipeRefreshLayout.OnRefreshListener;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.LayoutRes;
+import androidx.annotation.NonNull;
+import com.google.android.material.appbar.AppBarLayout;
+import com.google.android.material.appbar.AppBarLayout.LayoutParams;
+import com.google.android.material.appbar.AppBarLayout.LayoutParams.ScrollFlags;
+import androidx.coordinatorlayout.widget.CoordinatorLayout;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout.OnRefreshListener;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -54,21 +54,7 @@ import io.atlant.wallet.drawer_menu.DrawerHelper;
 
 import io.atlant.wallet.MyApplication;
 import atlant.wallet.R;
-import io.atlant.wallet.activities.backup.BackupActivity;
-import io.atlant.wallet.activities.camera.CameraActivity;
-import io.atlant.wallet.activities.export_wallet.ExportActivity;
-import io.atlant.wallet.activities.home.HomeActivity;
-import io.atlant.wallet.activities.import_wallet.ImportActivity;
-import io.atlant.wallet.activities.login.LoginActivity;
-import io.atlant.wallet.activities.login_selected.LoginSelectedActivity;
-import io.atlant.wallet.activities.login_selected_app.LoginSelectedAppActivity;
-import io.atlant.wallet.activities.main.MainActivity;
-import io.atlant.wallet.activities.receive.ReceiveActivity;
-import io.atlant.wallet.activities.send.SendActivity;
-import io.atlant.wallet.activities.settings.SettingsActivity;
 import io.atlant.wallet.dagger2.component.AppComponent;
-import io.atlant.wallet.drawer_menu.DrawerContent;
-import io.atlant.wallet.drawer_menu.DrawerHelper;
 import io.atlant.wallet.events.login.CredentialsCleared;
 import io.atlant.wallet.utils.CredentialHolder;
 import io.atlant.wallet.utils.DialogUtils;
