@@ -10,8 +10,8 @@ import android.content.ComponentName;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.view.ViewPager;
+import androidx.core.app.ActivityCompat;
+import androidx.viewpager.widget.ViewPager;
 import android.widget.TextView;
 import butterknife.BindView;
 import atlant.wallet.R;
