@@ -7,8 +7,8 @@
 package io.atlant.wallet.fragments.transactions;
 
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import butterknife.BindView;
 import atlant.wallet.R;
