@@ -9,8 +9,8 @@ package io.atlant.wallet.views;
 import android.app.Dialog;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.DrawableRes;
-import android.support.v4.content.ContextCompat;
+import androidx.annotation.DrawableRes;
+import androidx.core.content.ContextCompat;
 import android.view.View;
 import android.view.Window;
 import android.widget.ImageView;
